requirements.txt
opencv-python
python-dotenv
flask
PyJWT
ultralytics